initial.rev4
With e10s boostrap addons sometimes load before windows on application startup. Well I'm not sure if it's e10s or what but I have encountered bootstrap startup before windows, and on register it used to assume all windows are loaded. So now modified it so that on register it checks if loaded and acts appropriately.